requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dateparser==1.2.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logging==0.4.9.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.22.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
patsy==1.0.1
pillow==11.1.0
plotly==5.24.1
plotly-express==0.4.1
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.19.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
statsmodels==0.14.4
streamlit==1.41.1
ta==0.11.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
watchdog==6.0.0
websockets==14.1
yarl==1.18.3
